requirements.txt
playwright
requests
python-dotenv
flask
sqlalchemy
apscheduler
openai           # solo si usas la parte de IA, si no, puedes omitirla